# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "스터디 모집·신청 서비스",
    "description": "\n해커톤 연습용 스터디 모집·신청 서비스 API.\n\n**인증 없음**: 요청자는 `user_id`로만 구분한다.\n- `GET` / `DELETE` → `user_id`를 **쿼리**로 전달\n- `POST` / `PATCH` → `user_id`를 **body**로 전달\n- 방장 전용 API는 `user_id ≠ owner_id`이면 403\n\n에러 응답은 모두 `{\"detail\": \"한국어 메시지\"}` 형식이다. 시간은 KST naive datetime(`YYYY-MM-DDTHH:MM:SS`).\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/studies": {
      "post": {
        "tags": [
          "studies"
        ],
        "summary": "스터디 모집 글 생성",
        "description": "생성자는 자동으로 방장(OWNER) 멤버로 등록된다. 정원은 2 이상.",
        "operationId": "create_study_studies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "studies"
        ],
        "summary": "스터디 목록",
        "description": "`status`로 모집중(RECRUITING)/마감(CLOSED) 필터. 현재 인원수(`member_count`) 포함.",
        "operationId": "list_studies_studies_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StudyStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudyOut"
                  },
                  "title": "Response List Studies Studies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/studies/{study_id}": {
      "get": {
        "tags": [
          "studies"
        ],
        "summary": "스터디 상세",
        "description": "현재 인원수와 멤버 목록(방장 먼저)을 포함한다.",
        "operationId": "get_study_studies__study_id__get",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDetail"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/studies/{study_id}/capacity": {
      "patch": {
        "tags": [
          "studies"
        ],
        "summary": "정원 변경 (방장)",
        "description": "현재 인원보다 작으면 400.\n정원 > 인원이면 RECRUITING, 정원 == 인원이면 CLOSED(남은 대기 신청은 자동 거절)로 바뀐다.",
        "operationId": "update_capacity_studies__study_id__capacity_patch",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapacityUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyOut"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/studies/{study_id}/transfer-owner": {
      "post": {
        "tags": [
          "studies"
        ],
        "summary": "방장 위임 (방장)",
        "description": "대상은 해당 스터디 멤버여야 하고, 자기 자신은 400.\n한 트랜잭션에서 owner_id 변경 + 이전 방장 MEMBER, 새 방장 OWNER.",
        "operationId": "transfer_owner_studies__study_id__transfer_owner_post",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwner"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDetail"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/rooms/{room_id}/reservations": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "방의 날짜별 예약",
        "description": "해당 날짜에 걸쳐 있는 확정(CONFIRMED) 예약만 반환한다.",
        "operationId": "room_reservations_rooms__room_id__reservations_get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "YYYY-MM-DD",
              "title": "Date"
            },
            "description": "YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationOut"
                  },
                  "title": "Response Room Reservations Rooms  Room Id  Reservations Get"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/studies/{study_id}/reservations": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "스터디의 예약 목록",
        "description": "확정(CONFIRMED) 예약만 시작 시간순으로 반환한다.",
        "operationId": "study_reservations_studies__study_id__reservations_get",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationOut"
                  },
                  "title": "Response Study Reservations Studies  Study Id  Reservations Get"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "스터디룸 예약 (방장)",
        "description": "시간은 KST 기준 `YYYY-MM-DDTHH:MM:SS` (타임존이 붙으면 KST로 변환).\nend > start, 과거 시간 불가, 스터디 인원이 방 정원을 넘으면 400.\n같은 방의 확정 예약과 시간이 겹치면 409 (맞닿는 시간은 허용).",
        "operationId": "create_reservation_studies__study_id__reservations_post",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "중복/충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reservations/{reservation_id}": {
      "delete": {
        "tags": [
          "rooms"
        ],
        "summary": "예약 취소 (방장)",
        "description": "상태가 CANCELED로 바뀐다. 이미 취소된 예약이면 400.",
        "operationId": "cancel_reservation_reservations__reservation_id__delete",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reservation Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "요청자(방장) user_id",
              "title": "User Id"
            },
            "description": "요청자(방장) user_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationOut"
                }
              }
            }
          },
          "400": {
            "description": "규칙 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음 (방장/본인 아님)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapacityUpdate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "examples": [
              1
            ]
          },
          "capacity": {
            "type": "integer",
            "minimum": 2.0,
            "title": "Capacity",
            "examples": [
              5
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "capacity"
        ],
        "title": "CapacityUpdate"
      },
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "방장만 할 수 있는 작업입니다."
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MemberOut": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "nickname",
          "role"
        ],
        "title": "MemberOut"
      },
      "MemberRole": {
        "type": "string",
        "enum": [
          "OWNER",
          "MEMBER"
        ],
        "title": "MemberRole"
      },
...
      "ReservationCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "examples": [
              1
            ]
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id",
            "examples": [
              1
            ]
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At",
            "examples": [
              "2030-01-01T14:00:00"
            ]
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At",
            "examples": [
              "2030-01-01T16:00:00"
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "room_id",
          "start_at",
          "end_at"
        ],
        "title": "ReservationCreate"
      },
      "ReservationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "reserved_by": {
            "type": "integer",
            "title": "Reserved By"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "room_id",
          "study_id",
          "reserved_by",
          "start_at",
          "end_at",
          "status"
        ],
        "title": "ReservationOut"
      },
      "ReservationStatus": {
        "type": "string",
        "enum": [
          "CONFIRMED",
          "CANCELED"
        ],
        "title": "ReservationStatus"
      },
...
      "StudyCreate": {
        "properties": {
          "owner_id": {
            "type": "integer",
            "title": "Owner Id",
            "examples": [
              1
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "examples": [
              "알고리즘 스터디"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": "",
            "examples": [
              "주 2회 백준 문제 풀이"
            ]
          },
          "capacity": {
            "type": "integer",
            "minimum": 2.0,
            "title": "Capacity",
            "examples": [
              4
            ]
          }
        },
        "type": "object",
        "required": [
          "owner_id",
          "title",
          "capacity"
        ],
        "title": "StudyCreate"
      },
      "StudyDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "status": {
            "$ref": "#/components/schemas/StudyStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberOut"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "owner_id",
          "capacity",
          "status",
          "created_at",
          "member_count",
          "members"
        ],
        "title": "StudyDetail"
      },
      "StudyOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "status": {
            "$ref": "#/components/schemas/StudyStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "owner_id",
          "capacity",
          "status",
          "created_at",
          "member_count"
        ],
        "title": "StudyOut"
      },
      "StudyStatus": {
        "type": "string",
        "enum": [
          "RECRUITING",
          "CLOSED"
        ],
        "title": "StudyStatus"
      },
      "TransferOwner": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "examples": [
              1
            ]
          },
          "new_owner_id": {
            "type": "integer",
            "title": "New Owner Id",
            "examples": [
              2
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "new_owner_id"
        ],
        "title": "TransferOwner"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}